refactor rgdparser
