+++ CMakeLists.txt
@@ -35,6 +35,7 @@ add_subdirectory(compiler)
 add_subdirectory(instrumentation)
 add_subdirectory(runtime)
 add_subdirectory(wrappers)
+add_subdirectory(parsers)
 add_subdirectory(solvers)
 add_subdirectory(driver)
 add_subdirectory(tests)
